[rename] : 인증 로직 파일 구조 변경 (Member -> Auth)
